% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_mitogenome}
\alias{human_mitogenome}
\title{Human Mitochondrial Genome Sequence}
\format{
## `human_mitogenome`
Human Mitochondrial Genome Sequence
}
\usage{
human_mitogenome
}
\description{
Human Mitochondrial Genome Sequence
}
\keyword{datasets}
