% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign2base.R
\name{sign2base}
\alias{sign2base}
\title{Returns a DNA base based on the sign value of an X and Y coordinate}
\arguments{
\item{sx}{sign of X coordinate}

\item{sy}{sign of Y coordinate}
}
\value{
The corresponding base
}
\description{
The function takes the sign of an X and Y value and returns the corresponding DNA base.

Version 0.0.1.
Author: Dr. Matthew Cserhati
Email: matthew.cserhati@cui.edu
May 5, 2025
}
\examples{
sign2base(-1,1)
}
\references{
Jeffrey, H. J. (1990) Chaos game representation of gene structure. Nucleic Acids Research 18(8):2163-70.

Hill, K. A., Schisler, N. J., and Singh, S. M. (1992) Chaos game representation of coding regions of human globin genes and alcohol dehydrogenase genes of phylogenetically divergent species. Journal of Molecular Evolution 35:261-269.

Löchel, H. F., and Heider, D. (2021) Chaos game representation and its applications in bioinformatics. Computational and Structural Biotechnology Journal 19(2021): 6263-6271.
}
